extract usePerformanceReport into own hook and guard agains invalid urls being saved.
