+++ client/hosting/performance/hooks/usePerformanceReport.ts
@@ -1,0 +1,110 @@
+import { useState, useCallback, useEffect } from 'react';
+import { useUrlBasicMetricsQuery } from 'calypso/data/site-profiler/use-url-basic-metrics-query';
+import { useUrlPerformanceInsightsQuery } from 'calypso/data/site-profiler/use-url-performance-insights';
+import { TabType } from 'calypso/performance-profiler/components/header';
+
+function isValidURL( url: string ) {
+	return /^(https?:\/\/)?([a-z0-9-]+\.)+[a-z]{2,}(:[0-9]{1,5})?(\/[^\s]*)?$/i.test( url );
+}
+
+export const usePerformanceReport = (
+	setIsSavingPerformanceReportUrl: ( isSaving: boolean ) => void,
+	refetchPages: () => void,
+	savePerformanceReportUrl: (
+		pageId: string,
+		wpcom_performance_report_url: { url: string; hash: string }
+	) => Promise< void >,
+	currentPageId: string,
+	wpcom_performance_report_url: { url: string; hash: string } | undefined,
+	activeTab: TabType
+) => {
+	const { url = '', hash = '' } = wpcom_performance_report_url || {};
+
+	const [ retestState, setRetestState ] = useState( 'idle' );
+
+	const {
+		data: basicMetrics,
+		isError: isBasicMetricsError,
+		isFetched: isBasicMetricsFetched,
+		isLoading: isLoadingBasicMetrics,
+		refetch: requeueAdvancedMetrics,
+	} = useUrlBasicMetricsQuery( url, hash, true );
+
+	const { final_url: finalUrl, token } = basicMetrics || {};
+
+	useEffect( () => {
+		if ( token && finalUrl && isValidURL( finalUrl ) ) {
+			setIsSavingPerformanceReportUrl( true );
+			savePerformanceReportUrl( currentPageId, { url: finalUrl, hash: token } )
+				.then( () => {
+					refetchPages();
+				} )
+				.finally( () => {
+					setIsSavingPerformanceReportUrl( false );
+				} );
+		}
+		// We only want to run this effect when the token changes.
+		// eslint-disable-next-line react-hooks/exhaustive-deps
+	}, [ token ] );
+
+	const {
+		data,
+		status: insightsStatus,
+		isError: isInsightsError,
+		isLoading: isLoadingInsights,
+	} = useUrlPerformanceInsightsQuery( url, token ?? hash );
+
+	const performanceInsights = data?.pagespeed;
+
+	const mobileReport =
+		typeof performanceInsights?.mobile === 'string' ? undefined : performanceInsights?.mobile;
+	const desktopReport =
+		typeof performanceInsights?.desktop === 'string' ? undefined : performanceInsights?.desktop;
+
+	const performanceReport = activeTab === 'mobile' ? mobileReport : desktopReport;
+
+	const desktopLoaded = typeof performanceInsights?.desktop === 'object';
+	const mobileLoaded = typeof performanceInsights?.mobile === 'object';
+
+	const getHashOrToken = (
+		hash: string | undefined,
+		token: string | undefined,
+		isReportLoaded: boolean
+	) => {
+		if ( hash ) {
+			return hash;
+		} else if ( token && isReportLoaded ) {
+			return token;
+		}
+		return '';
+	};
+
+	const testAgain = useCallback( async () => {
+		setRetestState( 'queueing-advanced' );
+		const result = await requeueAdvancedMetrics();
+		setRetestState( 'polling-for-insights' );
+		return result;
+	}, [ requeueAdvancedMetrics ] );
+
+	if (
+		retestState === 'polling-for-insights' &&
+		insightsStatus === 'success' &&
+		( activeTab === 'mobile' ? mobileLoaded : desktopLoaded )
+	) {
+		setRetestState( 'idle' );
+	}
+
+	return {
+		performanceReport,
+		url: finalUrl ?? url,
+		hash: getHashOrToken( hash, token, activeTab === 'mobile' ? mobileLoaded : desktopLoaded ),
+		isLoading:
+			isLoadingBasicMetrics ||
+			isLoadingInsights ||
+			( activeTab === 'mobile' ? ! mobileLoaded : ! desktopLoaded ),
+		isError: isBasicMetricsError || isInsightsError,
+		isBasicMetricsFetched,
+		testAgain,
+		isRetesting: retestState !== 'idle',
+	};
+};
